Score display

+++ GMTK-2024/Assets/_Scripts/Scriptables/Item.cs
@@ -3,6 +3,6 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "NewItem", menuName = "ScriptableObjects/Items", order = 2)]
 public class Item : ScriptableObject {
   public string Name;
-  public float Weight;
+  public int Weight;
   public Sprite Image;
 }
